+++ projectmanagement-api/src/main/java/com/cyacompany/projectmanagement_api/service/TimeFactorService.java
@@ -1,9 +1,12 @@
 package com.cyacompany.projectmanagement_api.service;
 
+import com.cyacompany.projectmanagement_api.exception.ResourceNotFoundException;
 import com.cyacompany.projectmanagement_api.model.TimeFactor;
 import com.cyacompany.projectmanagement_api.repository.TimeFactorRepository;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.beans.Transient;
 import java.util.List;
 
 @Service
@@ -20,14 +23,19 @@ public class TimeFactorService {
   }
 
   public TimeFactor getById(Integer id) {
-    return repository.findById(id).orElse(null);
+    return repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("TimeFactor not found with id: " + id));
   }
 
+  @Transactional
   public TimeFactor save(TimeFactor timeFactor) {
     return repository.save(timeFactor);
   }
 
+  @Transactional
   public void deleteById(Integer id) {
+    if (!repository.existsById(id)) {
+      throw new ResourceNotFoundException("TimeFactor not found with id: " + id);
+    }
     repository.deleteById(id);
   }
 }
